Ensure the function under wrap is also extended
Prior to this change only the returned, wrapped function was extended
with the properties of the wrapped function. The function that is passed
to the wrapping function (this is the function that wraps the selection
and ensure any transition is carried forward) did not have any of the
properties of the wrapped function.

This change extends the function that is passed to the wrapping function
with the properties of the wrapped function.