Pass the wrapped function as the second argument
Prior to this change, wrap would pass the function to be wrapped as the
first argument to the wrapping function, second argument would be the
selection and the rest are passed trough from the calling function.

This change swaps the wrapped function with the selection, the selection
is reserved for the selection.